+++ src/main/preload.ts
@@ -72,6 +72,8 @@ const getFileStoreApi = <FileName extends StoreFileName>(
 }
 contextBridge.exposeInMainWorld('apiCommonStorage', getFileStoreApi('common'))
 contextBridge.exposeInMainWorld('apiUserNodesStorage', getFileStoreApi('userNodes'))
+contextBridge.exposeInMainWorld('apiChainStorage', getFileStoreApi('userChains'))
+contextBridge.exposeInMainWorld('apiAssetStorage', getFileStoreApi('userAssets'))
 contextBridge.exposeInMainWorld('apiPoolsStorage', getFileStoreApi('pools'))
 
 //