+++ src/renderer/services/storage/index.ts
@@ -1,5 +1,7 @@
 import * as common from './common'
 import * as pools from './pools'
+import * as userChains from './userChains'
+import * as userAssets from './userChainTokens'
 import * as userNodes from './userNodes'
 
-export { common, userNodes, pools }
+export { common, userNodes, pools, userChains, userAssets }