+++ src/renderer/services/thorchain/thornode.ts
@@ -43,7 +43,7 @@ import * as Rx from 'rxjs'
 import * as RxOp from 'rxjs/operators'
 
 import { AssetRuneNative } from '../../../shared/utils/asset'
-import { isEnabledChain } from '../../../shared/utils/chain'
+import { isSupportedChain } from '../../../shared/utils/chain'
 import { WalletType } from '../../../shared/wallet/types'
 import { ZERO_BASE_AMOUNT } from '../../const'
 import { THORCHAIN_DECIMAL } from '../../helpers/assetHelper'
@@ -77,7 +77,8 @@ import {
   LoanOpenQuote,
   BlockInformation,
   LoanCloseQuote,
-  LoanCloseQuoteLD
+  LoanCloseQuoteLD,
+  NodeStatusEnum
 } from './types'
 
 const height: number | undefined = undefined
@@ -136,7 +137,7 @@ export const createThornodeService$ = (network$: Network$, clientUrl$: ClientUrl
               A.filterMap(({ chain, address, ...rest }) =>
                 // validate chain
                 chain !== undefined &&
-                isEnabledChain(chain) &&
+                isSupportedChain(chain) &&
                 // address is required
                 !!address
                   ? O.some({ chain, address, ...rest })
@@ -310,7 +311,7 @@ export const createThornodeService$ = (network$: Network$, clientUrl$: ClientUrl
           address: node_address,
           bond: baseAmount(total_bond, THORCHAIN_DECIMAL),
           award: baseAmount(current_award, THORCHAIN_DECIMAL),
-          status,
+          status: status as NodeStatusEnum,
           bondProviders: {
             nodeOperatorFee: baseAmount(bond_providers.node_operator_fee, THORCHAIN_DECIMAL),
             providers: Array.isArray(bond_providers.providers)