+++ src/setupTests.ts
@@ -56,6 +56,8 @@ global.window.apiUrl = { ...mockApi.apiUrl }
 global.window.apiHDWallet = { ...mockApi.apiHDWallet }
 global.window.apiCommonStorage = { ...mockApi.apiCommonStorage }
 global.window.apiUserNodesStorage = { ...mockApi.apiUserNodesStorage }
+global.window.apiChainStorage = { ...mockApi.apiChainStorage }
+global.window.apiAssetStorage = { ...mockApi.apiAssetStorage }
 global.window.apiPoolsStorage = { ...mockApi.apiPoolsStorage }
 
 /**