+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.22.1 (2024-08-15)
+
+## Updates/Fixes
+- Bond tools update [#309](https://github.com/asgardex/asgardex-desktop/pull/315)
+- Mayachain synth warning [#311](https://github.com/asgardex/asgardex-desktop/pull/315)
+
+## Add
+
+- Chain enabled disable [#281](https://github.com/asgardex/asgardex-desktop/pull/312)
+- Added erc20 manual add [#313] (https://github.com/asgardex/asgardex-desktop/pull/314)
+
 # 1.22.0 (2024-07-29)
 
 ## Updates/Fixes